scripts/requirements.txt
inventree
requests
beautifulsoup4
Pillow
numpy
